+++ tests/similarity/test_molecular_similarity.py
@@ -43,6 +43,30 @@ def test_permutational_alignment2():
     assert np.all(permutation == np.array([1, 3, 2, 0, 5, 4]))
     assert np.all(perm_coords == pytest.approx(coords.position))
 
+def test_get_permutable_groups():
+    position = np.array([0.7430002202, 0.2647603899, -0.0468575389,
+                       -0.7430002647, -0.2647604843, 0.0468569750,
+                        0.1977276118, -0.4447220146, 0.6224700350,
+                        -0.1977281310, 0.4447221826, -0.6224697723,
+                        -0.1822009635, 0.5970484122, 0.4844363476,
+                         0.1822015272, -0.5970484858, -0.4844360463])
+    atom_labels = ['C','C','C','C','Ag','Ag']
+    coords = AtomicCoordinates(atom_labels, position)
+    similarity = MolecularSimilarity(0.1, 0.05)
+    coords_b = np.array([-0.1977281310, 0.4447221826, -0.6224697723,
+                         0.7430002202, 0.2647603899, -0.0468575389,
+                         0.1977276118, -0.4447220146, 0.6224700350,
+                        -0.7430002647, -0.2647604843, 0.0468569750,
+                         0.1822015272, -0.5970484858, -0.4844360463,
+                        -0.1822009635, 0.5970484122, 0.4844363476])
+    p1, p2 = similarity.get_permutable_groups(coords, coords_b)
+    if p1[0] == [4, 5]:
+        assert p1 == [[4, 5], [0, 1, 2, 3]]
+        assert p2 == [[4, 5], [0, 1, 2, 3]]
+    else:
+        assert p1 == [[0, 1, 2, 3], [4, 5]]
+        assert p2 == [[0, 1, 2, 3], [4, 5]]
+
 def test_random_rotation():
     position = np.array([0.7430002202, 0.2647603899, -0.0468575389,
                         -0.7430002647, -0.2647604843, 0.0468569750,
@@ -365,14 +389,14 @@ def test_permutational_alignment_mol():
     position = atoms.get_positions()
     coords = MolecularCoordinates(species, position.flatten())
     init_position = position.copy()
-    coords.position[0:3] = np.array([-1.2854, 0.2499, 0.0])
-    coords.position[3:6] = np.array([0.0072, -0.5687, 0.0])
+#    coords.position[0:3] = np.array([-1.2854, 0.2499, 0.0])
+#    coords.position[3:6] = np.array([0.0072, -0.5687, 0.0])
     coords.position[9:12] = np.array([-1.3175, 0.8784, 0.89])
     coords.position[15:18] = np.array([0.0392, -1.1972, 0.89])
     similarity = MolecularSimilarity(0.01, 0.05, weighted=False)
     p_coords, permutation = similarity.permutational_alignment(coords,
                                                                init_position.flatten())
-    assert np.all(permutation == np.array([1, 0, 2, 5, 4, 3, 6, 7, 8]))
+    assert np.all(permutation == np.array([0, 1, 2, 5, 4, 3, 6, 7, 8]))
     assert np.all(p_coords == pytest.approx(coords.position))
 
 def test_permutational_alignment_mol2():
@@ -381,17 +405,72 @@ def test_permutational_alignment_mol2():
     position = atoms.get_positions()
     coords = MolecularCoordinates(species, position.flatten())
     init_position = position.copy()
+#    coords.position[0:3] = np.array([-1.2854, 0.2499, 0.0])
+#    coords.position[3:6] = np.array([0.0072, -0.5687, 0.0])
[…]
