### 2024-04-25 #6580 @ml-evs
> 👋 @ml-evs - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I can review this but realistically not until w/c 13th May, feel free to pass over me if that is not quick enough
