### 2024-05-13 IBM/topography-searcher#3 @ml-evs
[JOSS REVIEW] Initial comments
I am raising this issue as part of the JOSS review over at https://github.com/openjournals/joss-reviews/issues/6711, and will use it to collect any comments that pop-up as I try out the code. This is from my first non-exhaustive pass of reviewing.

1. ~README has hardcoded version (0.0.2) on PyPI that does not match latest (0.0.3)~
2. ~The package does not have any continuous integration set up (the stub .travis.yml does nothing). This is not a requirement for JOSS but moves it down to "OK" from "Good" on this scale: https://joss.readthedocs.io/en/latest/review_criteria.html#tests. I would strongly recommend adding a simple GH actions CI if you want others to use and contribute to this code.~
3. ~After following the procedure in the README verbatim, 214 tests fail and 159 pass due to hardcoded paths. The package needs to be installed locally `pip install -e .` and the tests must be run from inside the tests directly only. Even after remedying these issues, there are still 5 test failures on my machine (details below) which seem to arise from naive FP comparisons:~

    <summary>Test failures (click to expand)</summary>

    ```
    ========================== short test summary info ==========================
    FAILED data/test_coordinates.py::test_get_bond_angle_info - assert [109.46888152...77610308, ...] == [109.46888152...77610308, ...]
    FAILED data/test_coordinates.py::test_get_bond_angle_info2 - assert [119.99846240...81713545, ...] == [119.99846240...81713545, ...]
    FAILED data/test_coordinates.py::test_get_bond_angle_info3 - assert [107.40877470...08213614, ...] == [107.40877470...08213614, ...]
    FAILED data/test_coordinates.py::test_get_specific_bond_angle_info - assert [109.46888152...77610308, ...] == [109.46888152...77610308, ...]
    FAILED data/test_coordinates.py::test_get_specific_bond_angle_info2 - assert [107.40877470...08213614, ...] == [107.40877470...08213614, ...]
    ====================== 5 failed, 357 passed in 48.55s =======================
    ```

4. ~There is limited documentation outside of the examples, which makes it hard to get an overview of the package and where to find certain functionality. From an initial cursory look it seems that most classes/methods have docstrings of varying usefulness, but ideally these would be combined into a searchable docs page using e.g., Sphinx or mkdocs. It is not a JOSS requirement that this is hosted online (though this is very useful) so a simple local docs build would suffice.~
5. ~I was unable to run `bayesian_optimisation.ipynb`, `dataset_roughness.ipynb` or `example_function.ipynb` due to an import error: `AttributeError: module 'topsearch' has no attribute 'potential'`. It looks like this module name was changed at some point. Following point 2) above, these examples should also be checked in the CI for each version.~
6. ~I tried the `scripts/atomic/generate_landscape/run.py` and my resulting `DisconnectivityGraph.png` does not match the one in `expected_output` at all (by my eye, at least, see below). Which should be trusted?~

    <summary>My disconnectivity graph output (click to expand)</summary>

    ![DisconnectivityGraph](https://github.com/IBM/topography-searcher/assets/7916000/60683d66-1bf2-424a-b072-cb890630a20f)
