### 2024-05-13 #6711 @ml-evs
I have made my initial comments at https://github.com/IBM/topography-searcher/issues/3 but ran into a few issues and am as of yet am unable to verify the functionality of the package. I have limited availability this week and will be entirely unavailable next week, but happy to test out any fixes when I return.
