### 2024-05-20 3da7878
Updating testing to remove hardcoded paths to data
