+++ tests/potentials/test_gaussian_process.py
@@ -5,10 +5,12 @@ import networkx as nx
 from topsearch.potentials.gaussian_process import GaussianProcess
 from topsearch.data.model_data import ModelData
 
+current_dir = os.path.dirname(os.path.dirname((os.path.realpath(__file__))))
+
 def test_function_rbf():
     # Fix theta for testing
-    model_data = ModelData(training_file='test_data/training_gp.txt',
-                           response_file='test_data/response_gp.txt')
+    model_data = ModelData(training_file=f'{current_dir}/test_data/training_gp.txt',
+                           response_file=f'{current_dir}/test_data/response_gp.txt')
     gp = GaussianProcess(model_data=model_data, kernel_choice='RBF',
                          kernel_bounds=[(1e1, 1.00000001e1),
                                         (1e1, 1.00000001e1),
@@ -27,8 +29,8 @@ def test_function_rbf():
     
 def test_function_and_std_rbf():
     # Fix theta for testing
-    model_data = ModelData(training_file='test_data/training_gp.txt',
-                           response_file='test_data/response_gp.txt')
+    model_data = ModelData(training_file=f'{current_dir}/test_data/training_gp.txt',
+                           response_file=f'{current_dir}/test_data/response_gp.txt')
     gp = GaussianProcess(model_data=model_data, kernel_choice='RBF',
                          kernel_bounds=[(1e1, 1.00000001e1),
                                         (1e1, 1.00000001e1),
@@ -49,8 +51,8 @@ def test_function_and_std_rbf():
 
 def test_function_rbf_fit():
     # Fit the rbf kernel
-    model_data = ModelData(training_file='test_data/training_gp.txt',
-                           response_file='test_data/response_gp.txt')
+    model_data = ModelData(training_file=f'{current_dir}/test_data/training_gp.txt',
+                           response_file=f'{current_dir}/test_data/response_gp.txt')
     gp = GaussianProcess(model_data=model_data, kernel_choice='RBF',
                          kernel_bounds=[(1e-1, 1e2), (1e-1, 1e2),
                                         (1e-5, 1e-4)],
@@ -71,8 +73,8 @@ def test_function_rbf_fit():
     assert gp_mean == pytest.approx(3.4859748905013)
 
 def test_standardise_training():
-    model_data = ModelData(training_file='test_data/training_gp.txt',
-                           response_file='test_data/response_gp.txt')
+    model_data = ModelData(training_file=f'{current_dir}/test_data/training_gp.txt',
+                           response_file=f'{current_dir}/test_data/response_gp.txt')
     gp = GaussianProcess(model_data=model_data, kernel_choice='RBF',
                          kernel_bounds=[(1e-1, 1e2), (1e-1, 1e2),
                                         (1e-5, 1e-4)],
@@ -88,8 +90,8 @@ def test_standardise_training():
     assert gp.get_score() == pytest.approx(0.9999999859966073)
 
 def test_standardise_response():
-    model_data = ModelData(training_file='test_data/training_gp.txt',
-                           response_file='test_data/response_gp.txt')
+    model_data = ModelData(training_file=f'{current_dir}/test_data/training_gp.txt',
+                           response_file=f'{current_dir}/test_data/response_gp.txt')
     gp = GaussianProcess(model_data=model_data, kernel_choice='RBF',
                          kernel_bounds=[(1e-1, 1e2), (1e-1, 1e2),
                                         (1e-5, 1e-4)],
@@ -103,8 +105,8 @@ def test_standardise_response():
     assert gp_mean == pytest.approx(-0.5768242790490604)
 
 def test_standardise_both():
-    model_data = ModelData(training_file='test_data/training_gp.txt',
-                           response_file='test_data/response_gp.txt')
[…]
