+++ tests/data/test_coordinates.py
@@ -4,9 +4,12 @@ import pytest
 import ase
 import numpy as np
 import os.path
+import os
 from topsearch.data.coordinates import StandardCoordinates, \
     AtomicCoordinates, MolecularCoordinates
 from topsearch.potentials.force_fields import MMFF94
+
+current_dir = os.path.dirname(os.path.dirname((os.path.realpath(__file__))))
 
 ########### TWO DIMENSIONS - EVEN BOUNDS ###############
 
@@ -355,7 +358,7 @@ def test_mol_initialisation():
                      [1, 6], [1, 7], [2, 8]]
 
 def test_get_bonds():
-    atoms = ase.io.read('test_data/ethanol.xyz')
+    atoms = ase.io.read(f'{current_dir}/test_data/ethanol.xyz')
     species = atoms.get_chemical_symbols()
     position = atoms.get_positions()
     coords = MolecularCoordinates(species, position.flatten())
@@ -367,7 +370,7 @@ def test_get_bonds():
                              [1, 6], [1, 7], [2, 8]]
 
 def test_get_bonds2():
-    atoms = ase.io.read('test_data/ethanol.xyz')
+    atoms = ase.io.read(f'{current_dir}/test_data/ethanol.xyz')
     species = atoms.get_chemical_symbols()
     position = atoms.get_positions()
     coords = MolecularCoordinates(species, position.flatten())
@@ -381,7 +384,7 @@ def test_get_bonds2():
                              [1, 2], [1, 3], [1, 4], [2, 8]]
 
 def test_get_bonds3():
-    atoms = ase.io.read('test_data/azobenzene.xyz')
+    atoms = ase.io.read(f'{current_dir}/test_data/azobenzene.xyz')
     species = atoms.get_chemical_symbols()
     position = atoms.get_positions()
     coords = MolecularCoordinates(species, position.flatten())
@@ -396,7 +399,7 @@ def test_get_bonds3():
 
 def test_get_bonds4():
     # dissociation
-    atoms = ase.io.read('test_data/ethanol.xyz')
+    atoms = ase.io.read(f'{current_dir}/test_data/ethanol.xyz')
     species = atoms.get_chemical_symbols()
     position = atoms.get_positions()
     coords = MolecularCoordinates(species, position.flatten())
@@ -410,7 +413,7 @@ def test_get_bonds4():
                              [1, 6], [1, 7], [2, 8]]
 
 def test_same_bonds():
-    atoms = ase.io.read('test_data/ethanol.xyz')
+    atoms = ase.io.read(f'{current_dir}/test_data/ethanol.xyz')
     species = atoms.get_chemical_symbols()
     position = atoms.get_positions()
     coords = MolecularCoordinates(species, position.flatten())
@@ -420,7 +423,7 @@ def test_same_bonds():
     assert same == True
 
 def test_same_bonds2():
-    atoms = ase.io.read('test_data/ethanol.xyz')
+    atoms = ase.io.read(f'{current_dir}/test_data/ethanol.xyz')
     species = atoms.get_chemical_symbols()
     position = atoms.get_positions()
     coords = MolecularCoordinates(species, position.flatten())
@@ -428,7 +431,7 @@ def test_same_bonds2():
     assert same == True
 
 def test_same_bonds3():
-    atoms = ase.io.read('test_data/ethanol.xyz')
+    atoms = ase.io.read(f'{current_dir}/test_data/ethanol.xyz')
     species = atoms.get_chemical_symbols()
     position = atoms.get_positions()
     coords = MolecularCoordinates(species, position.flatten())
@@ -438,7 +441,7 @@ def test_same_bonds3():
[…]
