### 2024-05-20 da7cd9a
Updating the example notebooks
