+++ .travis.yml
@@ -1,23 +1,8 @@
-language: ruby
-
-before_install:
-  - echo "#"
-  - echo "#"
-  - echo "TravisCI is unbelively powerful, but you need to do your research first."
-  - echo "#"
-  - echo "#"
-
+language: python
+python:
+   - "3.10"
+   - "3.11"
+install:
+   - pip install -r requirements.txt
 script:
-  - echo "#"
-  - echo "#"
-  - echo "Please take a look https://docs.travis-ci.com/user/tutorial/ for you options."
-  - echo "#"
-  - echo "#"
-
-
-after_success:
-  - echo "#"
-  - echo "#"
-  - echo "Don't forget to enable it in the GitHub repository also.."
-  - echo "#"
-  - echo "#"
+   - pytest
