### 2024-05-20 7f74082
Updating the atomic example
