+++ .travis.yml
@@ -1,8 +1,0 @@
-language: python
-python:
-   - "3.10"
-   - "3.11"
-install:
-   - pip install -r requirements.txt
-script:
-   - pytest
