+++ .github/workflows/python-app.yml
@@ -1,0 +1,19 @@
+name: topsearch
+
+on: [push]
+
+jobs:
+  build:
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+      - name: Test with pytest
+        run: |
+          pytest
