+++ .github/workflows/python-app.yml
@@ -17,7 +17,6 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
       - name: Test with pytest
         run: |
           pytest
