+++ .github/workflows/python-app.yml
@@ -17,6 +17,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install pytest
       - name: Test with pytest
         run: |
           pytest
