+++ .github/workflows/python-app.yml
@@ -18,6 +18,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install pytest
+          pip install pytest-cov
       - name: Test with pytest
         run: |
           pytest
