+++ .github/workflows/python-app.yml
@@ -16,9 +16,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
-          pip install pytest
-          pip install pytest-cov
+          python -m pip install
       - name: Test with pytest
         run: |
           pytest
