+++ .github/workflows/python-app.yml
@@ -16,7 +16,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          python -m pip install
+          python -m pip install --upgrade pip
       - name: Test with pytest
         run: |
           pytest
