+++ .github/workflows/python-app.yml
@@ -17,6 +17,8 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install -r requirements.txt
       - name: Test with pytest
         run: |
+          pip install pytest pytest-cov
           pytest
