+++ .github/workflows/python-app.yml
@@ -16,8 +16,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          python -m pip install -r requirements.txt
       - name: Test with pytest
         run: |
           pip install pytest pytest-cov
