+++ .github/workflows/python-app.yml
@@ -16,7 +16,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          python -m pip install -r requirements.txt
+          python -m pip install .
       - name: Test with pytest
         run: |
           pip install pytest pytest-cov
