### 2024-05-20 f659897
Update deprecated method in stationary_points
