+++ .github/workflows/python-app.yml
@@ -22,4 +22,5 @@ jobs:
           pip install pytest pytest-cov
           sudo apt install texlive
           sudo apt install dvipng texlive-latex-extra texlive-fonts-recommended
+          sudo apt install cm-super
           pytest
