### 2024-05-21 921a8ee
Updated docstrings and added static docs pages
