+++ src/topsearch/data/model_data.py
@@ -1,5 +1,5 @@
-""" Module that contains the ModelData class that stores and operates on
-    a dataset for use in Potential classes """
+""" Module that contains the ModelData class. This class stores and operates on
+    a training dataset for use in machine learning applications """
 
 import numpy as np
 from nptyping import NDArray
@@ -25,7 +25,7 @@ class ModelData:
     train_props : dict
         Dictionary containing the statistics of the training array
     hull : NDArray
-
+        Object storing the convex hull of the training data
     """
 
     def __init__(self,
@@ -50,7 +50,7 @@ class ModelData:
 
     def read_data(self, training_file: str, response_file: str) -> None:
         """ Read in the training and repsonse values needed for an ML model.
-            Stored in class attributes to be accessed through potential """
+            Stored in class attributes """
         self.training = np.loadtxt(f"{training_file}", dtype=float, ndmin=2)
         self.response = np.loadtxt(f"{response_file}", dtype=float)
         # Reset number of points and dimensions
