### 2024-05-21 #6711 @lbl59
[#6711](https://github.com/openjournals/joss-reviews/issues/6711) I ran into an error when I attempted to run block 12 in the `dataset_roughness.ipynb` notebook
<details><summary>See screenshot of the error</summary>
![image](https://github.com/openjournals/joss-reviews/assets/71844489/959674c3-7204-44b8-a208-5eb6d5be9263)

The issue was resolved after adding the following two lines in the code block prior to the main code:
```python
import matplotlib as mpl
mpl.rcParams.update(mpl.rcParamsDefault)
```

Is there any way for this fix to be built into the source code?
