### 2024-05-21 67f3102
Updating floating point comparison test window
