### 2024-05-21 7696cfe
Further changes to floating point comparison in testing
