### 2024-05-21 IBM/topography-searcher#3 @LukeDicks
@ml-evs thank you very much for the detailed feedback and recommendations, it’s much appreciated. We’ve made changes/added code and I’ve listed what we changed below, lining up with each of your previous points.
1.	Changed to `pip install topsearch` to ensure the most up-to-date version is installed.
2.	As recommended we have added Github actions CI. We now run the entire test framework for Python versions 3.10 and 3.11 on all pushes.
3.	We have modified the hard-coded paths to be relative, allowing pytest to be called from the root directory too without failures. We have also updated the tests to avoid failures due to naive floating point comparison, and these test successes are validated in the CI. Finally, we updated the installation instructions.
4.	We have generated documentation with Sphinx, which we have used to populate GitHub pages associated with the project. These pages include an index of (and search function for) all modules, along with their attributes and methods. These GH pages are now linked in the README.
5.	These notebooks were not correctly migrated to the open-source project, resulting in the many failures. We have now updated them so they work as intended. We have added a notebook for atomic clusters.
6.	These disconnectivity graphs are not too dissimilar in content despite their appearance. The results in expected_output are correct and we have updated parts of the script to ensure replication of these results.

I'll keep this issue open so you can continue to collect comments in one place.
