### 2024-05-22 #6711 @LukeDicks
@lbl59 Thanks for your feedback on the notebook. We've added that code block to `dataset_roughness.ipynb` and to the other notebooks too. We've also updated the default for that notebook to run in serial, which should remove the ValueError when running on Windows machines.
