### 2024-05-28 #6711 @LukeDicks
@crvernon happy to update. Both checklists are partially complete. @ml-evs has been through the code once and raised an issue on the Github repo listing problems and additional features that would be required. I’ve been through to edit and add these requests and listed my progress in that same issue. The issue has remained open to allow Matthew to check these updates once he’s available.
@lbl59 has been through the code and raised issues with the notebooks in this thread, which should be fixed now too.
