### 2024-06-11 #6711 @lbl59
Replying to @crvernon regarding [this comment](https://github.com/openjournals/joss-reviews/issues/6711#issuecomment-2158024829)!

My checklist is almost complete except for the 'Example usage' item as I am still currently running into the same error when running the example notebooks.

<details><summary>See example error message</summary>

`FileNotFoundError                         Traceback (most recent call last)
File c:\Python312\Lib\site-packages\matplotlib\texmanager.py:250, in TexManager._run_checked_subprocess(cls, command, tex, cwd)
    [249](file:///C:/Python312/Lib/site-packages/matplotlib/texmanager.py:249) try:
--> [250](file:///C:/Python312/Lib/site-packages/matplotlib/texmanager.py:250)     report = subprocess.check_output(
    [251](file:///C:/Python312/Lib/site-packages/matplotlib/texmanager.py:251)         command, cwd=cwd if cwd is not None else cls._texcache,
    [252](file:///C:/Python312/Lib/site-packages/matplotlib/texmanager.py:252)         stderr=subprocess.STDOUT)
    [253](file:///C:/Python312/Lib/site-packages/matplotlib/texmanager.py:253) except FileNotFoundError as exc:

File c:\Python312\Lib\subprocess.py:466, in check_output(timeout, *popenargs, **kwargs)
    [464](file:///C:/Python312/Lib/subprocess.py:464)     kwargs['input'] = empty
--> [466](file:///C:/Python312/Lib/subprocess.py:466) return run(*popenargs, stdout=PIPE, timeout=timeout, check=True,
    [467](file:///C:/Python312/Lib/subprocess.py:467)            **kwargs).stdout

File c:\Python312\Lib\subprocess.py:548, in run(input, capture_output, timeout, check, *popenargs, **kwargs)
    [546](file:///C:/Python312/Lib/subprocess.py:546)     kwargs['stderr'] = PIPE
--> [548](file:///C:/Python312/Lib/subprocess.py:548) with Popen(*popenargs, **kwargs) as process:
    [549](file:///C:/Python312/Lib/subprocess.py:549)     try:

File c:\Python312\Lib\subprocess.py:1026, in Popen.__init__(self, args, bufsize, executable, stdin, stdout, stderr, preexec_fn, close_fds, shell, cwd, env, universal_newlines, startupinfo, creationflags, restore_signals, start_new_session, pass_fds, user, group, extra_groups, encoding, errors, text, umask, pipesize, process_group)
   [1023](file:///C:/Python312/Lib/subprocess.py:1023)             self.stderr = io.TextIOWrapper(self.stderr,
   [1024](file:///C:/Python312/Lib/subprocess.py:1024)                     encoding=encoding, errors=errors)
-> [1026](file:///C:/Python312/Lib/subprocess.py:1026)     self._execute_child(args, executable, preexec_fn, close_fds,
   [1027](file:///C:/Python312/Lib/subprocess.py:1027)                         pass_fds, cwd, env,
   [1028](file:///C:/Python312/Lib/subprocess.py:1028)                         startupinfo, creationflags, shell,
...
   (...)
    [267](file:///C:/Python312/Lib/site-packages/matplotlib/texmanager.py:267)             exc=exc.output.decode('utf-8', 'backslashreplace'))
    [268](file:///C:/Python312/Lib/site-packages/matplotlib/texmanager.py:268)         ) from None

RuntimeError: Failed to process string with tex because latex could not be found`

I suspect this error is due to an incorrect installation of a package somewhere in my Python Environment, but I would appreciate feedback from @LukeDicks in case this is not the case.

Another small (and completely optional) suggestion would be include an example of the application of the use of this package to explain the output of a commonly-used ML model on a problem with readily available data (i.e. on Kaggle).

One last suggestion would be to add links to each example's Jupyter Notebook on the documentation `[Tutorials page](https://ibm.github.io/topography-searcher/tutorials.html)` to easier navigation of the user.
