### 2024-06-14 #6711 @LukeDicks
@lbl59 Thanks a lot for these recommendations. I think explaining a standard model from Kaggle would be very useful and I have added direct links to the notebooks at the tutorials page now.

I have updated the notebooks again because you raised a useful point that if a LaTeX installation is not available (or not within the PATH for the Jupyter notebook) then these fail. I have now changed them such that you can run them without LaTeX text rendering, which should mean they work with all installations. In all of them there are now two lines that can be commented in order to run without LaTeX text rendering, and these are highlighted with a comment above them. Hopefully with these two lines commented you can run the examples now. You can search for LaTeX in order to find the comments, but for convenience they're at:
atomic_cluster.ipynb - final code block
dataset_roughness.ipynb - the second code block within Illustrative feature pair
example_function.ipynb - final code block of Initialise classes
