### 2024-06-17 2ffa96f
added flexibility to MACE model specification
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
