### 2024-06-17 7b84ef1
small bugfix for MACE molecular basin-hopping
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
