### 2024-06-17 dd4f893
update remove atom clashes
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
