### 2024-06-18 1355aac
use psi4 internal optimisation
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
