### 2024-06-18 e447702
permit use of psi4 for optimisations in basin hopping
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
