### 2024-06-18 be0bca8
minimal working example
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
