### 2024-06-19 4e551cd
analytic hessian computation implemented
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
