### 2024-06-19 IBM/topography-searcher#3 @ml-evs
Thanks for addressing these comments, I've started crossing them out in my OP as I go through them. One minor point, the repository now seems very large, (100 MB for a fresh git clone). I think this is because the docs build has been committed directly to the repo (fine in principle, though I would warn you that it might be worth removing this and instead adding to the docs build to your CI, so that GH pages always has the latest copy (definitely non-blocking for the paper).

Minor comments:

- The pyproject.toml lists `pytest-coverage` (a weird defunct package), I think this should be `pytest-cov` instead.
- **Dependency management**: Now that ase 3.23 has been released, it is  installed automatically when doing a fresh install. The only incompatibility seems to be that `import ase` in some of the tests files no longer works, and should be `import ase.io`. As this is a library rather than an "app", you may want to consider how best to version dependencies with poetry (or otherwise) to avoid accidental breakages (though this one is of course extremely minor). One approach for catching such errors would be to use dependabot to occasionally check for updates compatible with your `pyproject.toml` (https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file), though again I don't consider this as breaking. Perhaps instead it would be nice just to mention what poetry is, and that the poetry.lock file are the tested versions of the deps (if that is true).
- Finally, you may also consider using git tags + GitHub releases to annotate the repository with the state at which the PyPI versions were released, this can be very helpful for debugging (and gives you a nice announcement for latest features that you can link to on GitHub). This can even be set up to automatically archive the state of the repository at that point on Zenodo, which is indeed a requirement of JOSS (post-acceptance).
