### 2024-06-19 IBM/topography-searcher#3 @ml-evs
Otherwise, I'm happy that my comments have been addressed -- the notebook examples are really nice and illustrate the package well. I'll return to the JOSS checklist and see how much more there is left to do for this review. I'll leave the issue open pending your response about my minor comments above, but feel free to close it.
