### 2024-06-19 #6711 @ml-evs
I am now happy to consider my review complete, and my initial comments have been resolved (pending a few minor details the authors may want to consider before the final JOSS archived version: https://github.com/IBM/topography-searcher/issues/3#issuecomment-2178343592).

I am very comfortable recommending this for acceptance in JOSS. TopSearch provides a very useful and high-quality interface to methods that will continue to become increasingly important, and I applaud the authors (and IBM) for making this available under a permissive license. Both the examples in the repository and the introduction in the paper are clear and concise. If forced to make a comment, it would be nice to see one of the beautiful landscapes from the examples/README as a figure directly in the paper, but I will leave that to the authors choice.
