### 2024-06-19 05e8370
make psi4 less chatty
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
