### 2024-06-19 98f97f7
added aimnet2 support
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
