### 2024-06-24 6aca70c
added more output options
Signed-off-by: MorrowChem <joemorrow117@gmail.com>
