+++ .github/workflows/documentation.yml
@@ -18,10 +18,10 @@ jobs:
         run: |
           sphinx-build docs _build
       - name: Deploy to GitHub Pages
-        uses: peaceiris/actions-gh-pages@v3
-        if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
-        with:
-          publish_branch: gh-pages
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: _build/
-          force_orphan: true
+        uses: actions/deploy-pages@v4
+#        if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
+#        with:
+#          publish_branch: gh-pages
+#          github_token: ${{ secrets.GITHUB_TOKEN }}
+#          publish_dir: _build/
+#          force_orphan: true
