+++ .github/workflows/documentation.yml
@@ -14,9 +14,10 @@ jobs:
       - name: Install dependencies
         run: |
           pip install sphinx sphinx_rtd_theme myst_parser
+          pip install topsearch
       - name: Sphinx build
         run: |
-          sphinx-build docs _build
+          sphinx-build docs/source _build
       - name: Deploy to GitHub Pages
         uses: actions/deploy-pages@v4
 #        if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
