+++ .github/workflows/documentation.yml
@@ -8,7 +8,6 @@ permissions:
 jobs:
   docs:
     runs-on: ubuntu-latest
-    python-version: ["3.11"]
     permissions:
         pages: write
         id-token: write   
