+++ .github/workflows/documentation.yml
@@ -7,6 +7,8 @@ permissions:
 
 jobs:
   docs:
+    environment:
+        name: github-pages
     runs-on: ubuntu-latest
     permissions:
         pages: write
