+++ .github/workflows/documentation.yml
@@ -31,6 +31,8 @@ jobs:
   deploy:
     runs-on: ubuntu-latest
     needs: docs
+    environment:
+        url: ${{ steps.deployment.outputs.page_url }}
     permissions:
         pages: write
         id-token: write
