### 2024-06-24 c11e6ce
Removing static html files to test for automation
