+++ docs/build/html/_sources/modules/similarity/similarity2.rst.txt
@@ -1,6 +1,0 @@
-Similarity
-===========
-
-.. automodule:: topsearch.similarity.similarity
-   :members:
-
