+++ docs/build/html/_sources/modules/sampling/sampling.rst.txt
@@ -1,6 +1,0 @@
-Sampling
-==========
-
-Module for coordinating other classes to provide a set of simple methods for exploring surfaces.
-
-- :doc:`exploration`
