+++ docs/build/html/_sources/modules/transition_states/nudged_elastic_band.rst.txt
@@ -1,5 +1,0 @@
-Nudged Elastic Band
-====================
-
-.. automodule:: topsearch.transition_states.nudged_elastic_band
-   :members:
