+++ docs/build/html/_sources/modules/plotting/stationary_points.rst.txt
@@ -1,5 +1,0 @@
-Stationary points
-==================
-
-.. automodule:: topsearch.plotting.stationary_points
-   :members:
