+++ docs/build/html/_sources/modules/data/coordinates.rst.txt
@@ -1,6 +1,0 @@
-Coordinates
-==============
-
-.. automodule:: topsearch.data.coordinates
-   :members:
-
