+++ docs/build/html/_sources/modules/data/kinetic_transition_network.rst.txt
@@ -1,6 +1,0 @@
-Kinetic transition network
-===========================
-
-.. automodule:: topsearch.data.kinetic_transition_network
-   :members:
-
