+++ docs/build/html/_sources/_autosummary/topsearch.data.rst.txt
@@ -1,6 +1,0 @@
-topsearch.data
-==============
-
-.. currentmodule:: topsearch
-
-.. automodule:: data
