+++ docs/build/html/_sources/_autosummary/topsearch.transition_states.rst.txt
@@ -1,6 +1,0 @@
-topsearch.transition\_states
-============================
-
-.. currentmodule:: topsearch
-
-.. automodule:: transition_states
