+++ docs/build/html/_sources/modules/potentials/dataset_fitting.rst.txt
@@ -1,6 +1,0 @@
-Dataset fitting
-================
-
-.. automodule:: topsearch.potentials.dataset_fitting
-   :members:
-
