+++ docs/build/html/_sources/modules/plotting/disconnectivity.rst.txt
@@ -1,6 +1,0 @@
-Disconnectivity graphs
-=======================
-
-.. automodule:: topsearch.plotting.disconnectivity
-   :members:
-
