+++ docs/build/html/_sources/modules/potentials/bayesian_optimisation.rst.txt
@@ -1,6 +1,0 @@
-Bayesian optimisation
-=====================
-
-.. automodule:: topsearch.potentials.bayesian_optimisation
-   :members:
-
