### 2024-06-24 091cb4e
Updating installation instructions
