### 2024-06-24 f0e9485
Updates to the testing to avoid Latex installation
