+++ .github/workflows/automated-testing.yml
@@ -1,6 +1,6 @@
 name: automated-testing
 
-on: [push]
+on: [push, pull_request]
 
 jobs:
   build:
@@ -20,7 +20,7 @@ jobs:
       - name: Test with pytest
         run: |
           pip install pytest pytest-cov
-          sudo apt install texlive
-          sudo apt install dvipng texlive-latex-extra texlive-fonts-recommended
-          sudo apt install cm-super
+#          sudo apt install texlive
+#          sudo apt install dvipng texlive-latex-extra texlive-fonts-recommended
+#          sudo apt install cm-super
           pytest
