### 2024-06-24 f541283
Remove Latex from CI testing
