+++ .github/workflows/automated-testing.yml
@@ -20,7 +20,4 @@ jobs:
       - name: Test with pytest
         run: |
           pip install pytest pytest-cov
-#          sudo apt install texlive
-#          sudo apt install dvipng texlive-latex-extra texlive-fonts-recommended
-#          sudo apt install cm-super
           pytest
