### 2024-06-24 8b0b113
Avoid latex in testing
