### 2024-06-24 3c999cf
Plotting tests
