### 2024-06-24 eb6b92c
All plotting tests without Latex needed
