+++ tests/plotting/test_network.py
@@ -3,6 +3,8 @@ import numpy as np
 import os.path
 from topsearch.data.kinetic_transition_network import KineticTransitionNetwork
 from topsearch.plotting.network import plot_network, barrier_reweighting
+import matplotlib as mpl
+mpl.rcParams.update(mpl.rcParamsDefault)
 
 current_dir = os.path.dirname(os.path.dirname((os.path.realpath(__file__))))
 
@@ -27,6 +29,7 @@ def test_plot_network():
     ktn = KineticTransitionNetwork()
     ktn.read_network(text_path=f'{current_dir}/test_data/',
                      text_string='.network')
+    mpl.rcParams.update(mpl.rcParamsDefault)
     plot_network(ktn)
     assert os.path.exists('Network.png') == True
     os.remove('Network.png')
