### 2024-06-24 c53d18d
Removing unnecessary pdf of JOSS submission
