+++ joss_submission/paper.md
@@ -1,65 +1,0 @@
----
-title: 'TopSearch: a Python package for topographical analysis of machine learning models and physical systems'
-tags:
-  - Python
-  - machine learning
-  - topography
-  - energy landscapes
-  - chemical physics
-authors:
-  - name: Luke Dicks
-    orcid: 0000-0002-5278-4412
-    affiliation: "1"
-  - name: Edward O. Pyzer-Knapp
-    corresponding: true
-    orcid: 0000-0002-8232-8282
-    affiliation: 1
-affiliations:
- - name: IBM Research Europe, Hartree Centre, Daresbury, United Kingdom
-   index: 1
-date: 2 April 2024
-bibliography: paper.bib
----
-
-# Summary
-
-Machine learning (ML) is now ubiquitous in all scientific fields, but there remains a significant challenge to understanding and explaining model performance [@Zhang2021; @Angelov2021]. Therefore, there is increasing interest in applying methods from other scientific disciplines (e.g. physics and biology) to improve the performance and explainability of machine learning algorithms [@Hassabis2017; @Karniadakis2021]. One methodology that has proved useful to understand machine learning performance is the energy landscape framework from chemical physics [@Wales:2003].
-
-The energy landscape framework is a set of algorithms that map the topography of continuous surfaces by their stationary points. The topography is encoded as a weighted graph [@Noe2008] and in application to potential energy surfaces all physical properties of a system can be extracted from this graph [@Swinburne2020]. Examples of the methodology applied to potential energy surfaces explain physical phenomena for proteins [@Roder2019], small molecules [@Matysik2021], atomic clusters [@Csanyi2023] and crystalline solids [@Pracht2023].
-
-Since the energy landscape framework is applicable to any given continuous surface, the methodology can also be applied to a wide range of machine learning algorithms through the corresponding loss function surface. Fitting of a machine learning model usually aims to locate low-valued or diverse solutions, and an understanding of the solution space topography explains model reproducibility and performance. Leveraging the energy landscape framework the performance and reliability of neural networks [@Niroomand2022], Gaussian processes [@Niroomand:2023] and clustering algorithms [@Dicks2022; @Dicks2023; @Wu2023] has been explored. Moreover, it has been used to explain the effect of dataset roughness on ML model performance [@Dicks2024]. A tutorial review of different applications is given in @Niroomand:2024.
-
-# Statement of need
-
-The `topsearch` Python package provides a rapid prototyping software for application of the energy landscape framework. It contains the functionality to be used for both potential energy surfaces and the loss function surfaces of varied machine learning models.
-
-There is limited software for explicitly analysing the topography of loss function surfaces. These surfaces are considered implicitly when optimising an ML model through local minimisation, but none attempt to capture global topographical features of the parameter space. There is significantly more software for analysing potential energy surfaces, the majority of which approximate topographical features indirectly. Popular examples that aim to explore diverse regions of the surface through enhanced sampling are PyEMMA [@scherer_pyemma_2015] and large molecular simulation suites such as LAMMPS [@LAMMPS], GROMACS [@GROMACS], and AMBER [@AMBER] the simulations of which can be simplified using PLUMED [@PLUMED]. Explicit location of topographical features, such as stationary points, is more common in quantum chemistry and can be performed by software such as VTST [@vtst], PASTA [@Kundu2018], PyMCD [@Lee2023] and ORCA [@Neese2020]. The explicit computation of topography using the energy landscape framework has several advantages for application to machine learning and none of the above software contains all the required functionality.
-
[…]
