+++ joss_submission/paper.bib
@@ -1,283 +1,0 @@
-@book{Wales:2003,
-Author = {{Wales}, D.~J.},
-Publisher = {Cambridge University Press},
-Title = {{Energy Landscapes}},
-Year = 2003,
-doi={10.1017/CBO9780511721724},
-}
-
-@article{Niroomand:2023,
-    author = {M. P. Niroomand and L. Dicks and E. O. Pyzer-Knapp and D. J. Wales},
-    title = "{Physics inspired approaches to understanding Gaussian processes}",
-    journal = {arXiv},
-    archivePrefix = "arXiv",
-    year = 2023,
-    doi = {10.48550/arXiv.2305.10748},
-}
-
-@article{Niroomand2022,
-author = {M. P. Niroomand and C. T. Cafolla and J. W. R. Morgan and D. J. Wales},
-title = {Characterising the area under the curve loss function landscape},
-journal = {Mach. Learn.: Sci. Tech.},
-volume = {3},
-pages = {015019},
-year = {2022},
-doi = {10.1088/2632-2153/ac49a9},
-}
-
-@article{Dicks2022, 
-	author = {L. Dicks and D. J. Wales}, 
-	title = {{Elucidating the solution structure of the $K$-means cost function using energy landscape theory}},  
-	year = {2022}, 
-	journal = {J. Chem. Phys.},
-    pages   = {054109},
-    volume  = {156},
-    doi = {10.1063/5.0078793},
-}
-
-@article{Noe2008,
-  author  = {F. No{\'{e}} and S. Fischer}, 
-  title  =  {Transition networks for modelling the kinetics of conformational change in macromolecules},
-  journal = {Curr. Opin. Struct. Biol.},
-  year    = 2008,
-  volume  = 18,
-  pages   = {154-162},
-  doi = {10.1016/j.sbi.2008.01.008},
-}
-
-@article{Niroomand:2024,
-    author = {{Niroomand}, M.~P. and {Dicks}, L. and {Pyzer-Knapp}, E.~O. and {Wales}, D.~J.},
-    title = "{Insights into machine learning models from chemical physics: an energy landscapes approach (EL for ML)}",
-    journal = {Digital Discovery},
-    year = 2024,
-    doi = {10.1039/D3DD00204G},
-}
-
-@article{Dicks2023, 
-	author = {L. Dicks and D. J. Wales}, 
-	title = {{Evolution of $K$-means solution landscapes with the addition of dataset outliers and a robust clustering comparison measure for their analysis}},  
-	year = {2023}, 
-	journal = {arXiv},
-        doi = {10.48550/arXiv.2306.14346},
-}
-
-@article{Wu2023, 
-	author = {Y. Wu and L. Dicks and D. J. Wales}, 
-	title = {{Archetypal solution spaces for clustering gene expression datasets in identification of cancer subtypes}},  
-	year = {2023}, 
-	journal = {arXiv},
-        doi = {10.48550/arXiv.2305.17279},
-}
-
-@article{Dicks2024, 
-	author = {L. Dicks and D. E. Graff and K. E. Jordan and C. W. Coley and E. O. Pyzer-Knapp}, 
-	title = {{A physics-inspired approach to the understanding of molecular representations and models}},  
-	year = {2024}, 
-	journal = {Mol. Syst. Des. Eng.},
-        doi = {10.1039/D3ME00189J},
-}
-
-@misc{gmin,
-  author = {{D. J. Wales}},
-  title = {{GMIN}: a program for basin-hopping global optimisation, basin-sampling, and parallel tempering},
-  url = {https://www-wales.ch.cam.ac.uk/GMIN/},
[…]
