+++ pyproject.toml
@@ -16,7 +16,7 @@ scipy = "^1.11.3"
 numpy = "^1.26.1"
 scikit-learn = "^1.3.2"
 nptyping = "^2.5.0"
-ase = "^3.22.1"
+ase = "3.22.1"
 rdkit = "^2023.9.4"
 
 [tool.poetry.group.dev.dependencies]
