### 2024-06-24 IBM/topography-searcher#3 @LukeDicks
@ml-evs thank you so much for the further feedback, it's been very helpful.

- I have now removed the static html pages and added documentation generation into the CI in order to pull documentation for the current state of the code.

- You're right, I've changed it to `pytest-cov`.

- We've pinned the ase version now, and provided a short explanation of poetry as an alternative dependency manager, and that poetry.lock provides the specific tested version of those dependencies in the README and GitHub-pages.

- We've created a tagged release for the current v0.0.3 now. We will definitely do this going forwards for further releases.

Thanks again, I'll mark this issue as closed now.
