### 2024-06-30 #6711 @lbl59
@LukeDicks thank you for making the latex fixes and all your effort on the new notebook! @crvernon happy to declare my checklist done and all requests completed!
