### 2024-07-01 #6711 @crvernon
👋  @LukeDicks - we are almost there! Next is just setting up the archive for your new release.

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) if you have not already done so.  **Please ensure that the software archive uses the same license as the license you have posted on GitHub.**

- [ ]  Please respond with the released version number and DOI of the archived version here
