### 2024-07-01 34651b2
Update CITATION.cff
