+++ CITATION.cff
@@ -3,7 +3,8 @@ message: "If you use this software, please cite it as below."
 authors:
 - family-names: "Dicks"
   given-names: "Luke"
+  orcid: "https://orcid.org/0000-0002-5278-4412"
 title: "TopSearch: a Python package for exploring high-dimensional surfaces"
-version: 0.0.2
-date-released: 2024-22-03
+version: 0.0.3
+date-released: 2024-07-01
 url: "https://github.com/IBM/topography-searcher"
