### 2024-07-01 eb0f1b8
Update CITATION.cff
