### 2024-07-01 b8fa45b
Update CITATION.cff
