### 2024-07-01 54ba581
Delete CITATION.cff
