### 2024-07-01 7983ba7
Create CITATION.cff
