+++ CITATION.cff
@@ -1,0 +1,10 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: "Dicks"
+  given-names: "Luke"
+  orcid: "https://orcid.org/0000-0002-5278-4412"
+title: "TopSearch"
+version: 0.0.3
+date-released: 2024-07-01
+url: "https://github.com/IBM/topography-searcher"
