### 2024-07-02 #6711 @LukeDicks
@crvernon I have released version 0.0.3, and the doi for the archive of this on Zenodo is https://doi.org/10.5281/zenodo.12613017
