### 2024-07-02 #6711 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
